Hommexx: Work around a Cuda issue on Perlmutter GPU.
After an upstream merge to SCREAM, we encountered issue
https://github.com/E3SM-Project/scream/issues/2146. Some perturbation to the
code made something on the Cuda side exhibit a bug related to extended lambdas.
This commit works around the bug by moving the code to a functor to avoid the
lambda part of the call stack. To be clear, we have a bunch of lambdas; it's
just this one that is a problem.

Also remove an unused register_field overload.